requirements.txt
pip
numpy
pandas
sklearn
imblearn
tensorflow-gpu
tensorflow_addons
matplotlib
tqdm
